requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
coloredlogs==15.0.1
discord-ext-voice_recv==0.5.2a179
discord.py==2.5.2
flatbuffers==25.2.10
frozenlist==1.7.0
humanfriendly==10.0
idna==3.10
joblib==1.5.1
mpmath==1.3.0
multidict==6.5.0
numpy==1.26.4
onnxruntime==1.19.2
openwakeword==0.6.0
packaging==25.0
propcache==0.3.2
protobuf==6.31.1
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1
sounddevice==0.5.2
sympy==1.14.0
tflite-runtime==2.14.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.1
